+++ config.yaml
@@ -18,7 +18,7 @@ prune_params:
   er_method: just dont 
   prune_method: synflow
 experiment_params:
-  total_epochs: 90 
+  epochs_per_level: 90 
   num_levels: 30
   training_type: wr
 optimizer:
